+++ easypermissions-ktx/src/main/java/com/mvalceleanu/easypermissions/annotations/AfterPermissionGranted.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package mvalceleanu.easypermissions.annotations
+package com.mvalceleanu.easypermissions.annotations
 
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.FUNCTION)
