+++ easypermissions-ktx/src/test/java/pub/devrel/easypermissions/components/TestAppCompatActivity.kt
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package mvalceleanu.easypermissions.components
+package com.mvalceleanu.easypermissions.components
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import mvalceleanu.easypermissions.R
-import mvalceleanu.easypermissions.annotations.AfterPermissionGranted
-import mvalceleanu.easypermissions.annotations.Mockable
-import mvalceleanu.easypermissions.EasyPermissions
+import com.mvalceleanu.easypermissions.R
+import com.mvalceleanu.easypermissions.annotations.AfterPermissionGranted
+import com.mvalceleanu.easypermissions.annotations.Mockable
+import com.mvalceleanu.easypermissions.EasyPermissions
 
 @Mockable
 class TestAppCompatActivity :
